Fix BA's feedback (part 3) (#151)
* add modal

* edit modal

* fix carousel

* add w-23

* fix BA feedback for activity pages (not create new popup yet)

* created user act, fix carousel

* change w-25

* create RequireSignup component, change layout of Signup component

* fix boder Nav, hover rank

* fix boder nav

* fix font-size, button

* Remove right-rail of register page

* Temp cheat

* Revert appsettings.json

* add hover ability to subskill popup, delete devider at Act Info page

* fix BA's feedback (p1)

* Fix BA's feedback for create/edit page (p2)

* fix feedback, add user view

* add LoginRequire

* resize img

* fix navMenu,  carousel edit-org , play number,

* fix all conflict

* fix some stupid bug

* always paste without format for TinyMCE

* fix  access Org Profile link

* refactor code and add link for act item

* fix page edit user profile

* Minor fix

* link org page to place related

* add check infor user

* Re-structure folders

* Fix navbar

* Move SearchOrganizerPage out

* Fix route names and handle entry logic for org profile page

* Fix region

* fix play number, nav active

* fix data

* fix items in activity card

* add validation to check valid phonenumber

* Fix cannot edit org profile

* Fix img folders

* fix  popup edit user profile

* fix validation phonenumber

* fix logic signup popup.

* css for error-message at page edit user profile

* undo comment line 2 at page edit user profile

* fix img's height, z-index

* fix page of Khang, profile-org, number Homepage

* fix position button actCart

* fix backend logic to reuse Act component for homepage

* fix skillPopup

* fix volunteers list page

* reuse popup from org profile

* fix logic for reuse close/delete component

* add more button

* fix button's position

* fix modal "mo-hoat-dong" and "xoa-hoat-dong"

* delete popup

* fix dropdown list's position

* fix đại lộ danh vọng, fix org-profile, fix line, fix....

* change potision and color of validation phone number

* fix xong, còn lại: chưa resue act vào Homepage, chưa chỉnh các modal cho đồng nhất, carousel nhấn button còn chạy ngược

* create  ProfileRank component

* change take 4 to take 8 activities in homepage

* fix bg-color, đại lộ danh vọng

* fix css for AcitivityLogPage

* remove spinner

* Add AntDesign and implement datepicker with VN format

* Use Ant DatePicker for edit user profile and create activitiy pages

* fix onclick hide show sub skills

* fix phonenumber validation

* fix carousel, nav, footer, act

* fix chọn lĩnh vực

* fix area and deparment in edit user profile page

* fix UI areas and skills popup

* fix backend (faculty & area) - edit user profile

* fix logic for get organizer by include, not use identiy service anymore

* hot fix paging when filter or search activities

* Refactor and re-use filter logic from Activity Search page for 2 activity sections on Homepage

* refactor code

* [Backend] Handle logic for user profile page (#154)

* refactor code org service

* fix logic for access by route without orgid

* fix logic for calculate total rating and total rank of org

* handle upload avatar for user profile and remove uneccessary images

* fix profile rank component

* fix logic to show stars and total rating

* handle logic for show activities

* short cut code

* refactor code

* fix bug carrousel

* handle logic for appellation (unofficial)

* fix logic for show heartbreaker

* handle logic for show schedule activties

* fix logic schedule

* Handle backend logic for calendar

* fix startdate, enddate when create/update activity just take date, not time

* hot fix logic for get activitydaysasync

* hot fix after get latest code


* fix name of prop mistake

* fix some stuff

* Minor fix rightrail act info page

* Minor fixes img on act info page

* [Backend] Handle logic for pinned area (#146)

* handle logic for pinned area

* add color to area

* Handle area color in other places

* fix mapping IsPinned


* Add edit request card and create edit request page

* fix UI for EditRequestPage

* [Backend] Handle logic for open/close date activity (#162)

* big change logic to add open and close activity

* fix display signup button in activity search page

* fix logic for filter org act

* refactor method GetOrgActs

* fix logic for open/close date in org profile

* fix logic for open/close date in user profile

* fix some stuff

* fix logic for not show activity deleted in org act management

* big change logic to add open and close activity

* fix display signup button in activity search page

* fix logic for filter org act

* refactor method GetOrgActs

* fix logic for open/close date in org profile

* fix logic for open/close date in user profile

* fix some stuff

* fix logic for not show activity deleted in org act management

* rebase branch

* remove images

* refactor method GetAllActivitiesAsync

* move get all org from identity service to organization service

* fix background job hangfire

* fix logic, use DateTime.Date to make sure there is no mistake.

* fix logic calendar

* change logic PopUpReport to reuse it

* handle logic to report user and fix some stuff

* handle todo note

* fix logic to show button sign up

* fix logic to show button signup again

* remove navigation

* fix logic calendar

* fix chose area

+++ VMS/Pages/Organization/Profile/EditOrganizationProfile.razor.cs
@@ -11,6 +11,7 @@ using VMS.Application.Interfaces;
 using VMS.Application.ViewModels;
 using Microsoft.Extensions.Logging;
 using VMS.Common;
+using VMS.Common.Enums;
 
 namespace VMS.Pages.Organization.Profile
 {
@@ -33,20 +34,23 @@ namespace VMS.Pages.Organization.Profile
         [Inject]
         private ILogger<EditOrganizationProfile> Logger { get; set; }
 
+        [Inject]
+        private NavigationManager NavigationManager { get; set; }
+
         private int width;
         private string classWidth = "";
-        private bool isLoading;
         private string OrgId;
         private int count;
         private bool isErrorMessageShown = false;
         private IBrowserFile uploadFile;
         private IList<AreaViewModel> choosenAreas = new List<AreaViewModel>();
-        private CreateUserProfileViewModel org = new();
+        private CreateOrgProfileViewModel org = new();
+        [CascadingParameter] public string UserId { get; set; }
 
         protected override async Task OnInitializedAsync()
         {
             OrgId = IdentityService.GetCurrentUserId();
-            org = await UserService.GetCreateUserProfileViewModelAsync(OrgId);
+            org = await UserService.GetOrgProfileViewModelAsync(OrgId);
             choosenAreas = org.Areas;
         }
 
@@ -84,12 +88,11 @@ namespace VMS.Pages.Organization.Profile
             await areasModal.Result;
         }
 
+
+
         protected override async Task OnAfterRenderAsync(bool firstRender)
         {
-            if (choosenAreas.Count > 3)
-            {
-                await JSRuntime.InvokeVoidAsync("vms.EditProfileCarousel");
-            }
+            await JSRuntime.InvokeVoidAsync("vms.EditProfileCarousel", choosenAreas.Count);
         }
 
         private async Task HandleSubmit()
@@ -100,33 +103,46 @@ namespace VMS.Pages.Organization.Profile
                 return;
             }
 
-            Logger.LogInformation("HandleValidSubmit called");
-            isErrorMessageShown = false;
-            isLoading = true;
-
-            try
+            ModalOptions options = new()
             {
-                if (uploadFile is not null)
+                HideCloseButton = true,
+                DisableBackgroundCancel = true,
+                UseCustomLayout = true
+            };
+
+            IModalReference editConfirmModal = Modal.Show<EditConfirm>("", options);
+            ModalResult result = await editConfirmModal.Result;
+            if (!result.Cancelled)
+            {
+                Logger.LogInformation("HandleValidSubmit called");
+                isErrorMessageShown = false;
+
+                try
                 {
-                    string oldImageName = org.Banner;
-                    org.Banner = await UploadService.SaveImageAsync(uploadFile, OrgId);
-                    UploadService.RemoveImage(oldImageName);
-                }
-                await UserService.UpdateUserProfile(org, OrgId);
-                isLoading = false;
-            }
-            catch (Exception ex)
-            {
-                isLoading = false;
-                Logger.LogError("Error occurs when trying to edit profile", ex.Message);
-                await JSRuntime.InvokeVoidAsync("alert", ex.Message);
-            }
+                    if (uploadFile is not null)
+                    {
+                        string oldImageName = org.Banner;
+                        org.Banner = await UploadService.SaveImageAsync(uploadFile, OrgId, ImgFolder.Banner);
+                        UploadService.RemoveImage(oldImageName);
+                    }
 
+                    await UserService.UpdateOrgProfile(org, OrgId);
+
+                    ModalParameters modalParams = new();
+                    modalParams.Add("Title", succeededCreateTitle);
+                    await Modal.Show<Activities.NotificationPopup>("", modalParams, options).Result;
+                    NavigationManager.NavigateTo($"{Routes.OrgProfile}/{UserId}", true);
+                }
+                catch (Exception ex)
+                {
+                    Logger.LogError("Error occurs when trying to edit profile", ex.Message);
+                    await JSRuntime.InvokeVoidAsync("alert", ex.Message);
+                }
+            }
         }
 
         private async Task HandleInvalidSubmit()
         {
-            isLoading = false;
             isErrorMessageShown = true;
             await Interop.ScrollToTop(JSRuntime);
         }
